feat(notifications): moved notifications.dart into a service folder
